+++ src/components/Thanks/ThanksPageSingleVersion.vue
@@ -22,6 +22,32 @@
 				@continue-clicked="handleContinue(NON_TIERED_BADGE)"
 				class="tw-mb-2.5"
 			/>
+			<!--  Start of donation context -->
+			<OptInModule
+				v-if="showDonationOptInModule"
+				data-testid="donation-opt-in-module"
+				:loans="loans"
+				:is-guest="isGuest"
+				:number-of-badges="numberOfBadges"
+				:only-donations="true"
+				:achievements-completed="achievementsCompleted"
+				class="print:tw-hidden tw-mb-2.5"
+			/>
+			<AccountReceiptShare
+				v-if="onlyDonations"
+				ref="receiptSection"
+				:is-guest="isGuest"
+				:number-of-badges="numberOfBadges"
+				:receipt="receipt"
+				:lender="lender"
+				:loans="loans"
+				:show-receipt="showReceipt"
+				:only-donations="onlyDonations"
+				:guest-username="guestUsername"
+				class="tw-mb-2.5"
+			/>
+			<!--  End of donation context -->
+
 			<!-- Start goal module variations -->
 			<GoalEntrypoint
 				v-if="!isGuest && goalDataInitialized && isEmptyGoal"
@@ -59,12 +85,11 @@
 			/>
 			<!-- End goal module variations -->
 			<OptInModule
-				v-if="showOptInModule"
+				v-if="showNonDonationOptInModule"
 				data-testid="opt-in-module"
 				:loans="loans"
 				:is-guest="isGuest"
 				:number-of-badges="numberOfBadges"
-				:only-donations="onlyDonations"
 				:achievements-completed="achievementsCompleted"
 				class="print:tw-hidden tw-mb-2.5"
 			/>
@@ -99,6 +124,7 @@
 				@guest-continue="handleContinue"
 			/>
 			<AccountReceiptShare
+				v-if="!onlyDonations"
 				ref="receiptSection"
 				:is-guest="isGuest"
 				:number-of-badges="numberOfBadges"
@@ -226,6 +252,10 @@ const props = defineProps({
 		type: Array,
 		default: () => ([]),
 	},
+	goalRecommendedLoanEnable: {
+		type: Boolean,
+		default: false,
+	},
 });
 
 const badgeAchievedIds = ref(props.badgesAchieved.map(b => b.achievementId));
@@ -294,6 +324,8 @@ const hasTeamAttributedPartnerLoan = computed(
 );
 
 const showOptInModule = computed(() => !props.isOptedIn);
+const showDonationOptInModule = computed(() => showOptInModule.value && onlyDonations.value);
+const showNonDonationOptInModule = computed(() => showOptInModule.value && !onlyDonations.value);
 const showKivaCardsModule = computed(() => !!printableKivaCards.value.length);
 const showGoalCompletedModule = computed(() => {
 	// Show goal completed module immediately when user achieved their goal